+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   * File parsing with XTherion Insert Image Config being treated as if the last image set doesn't exist.
   * Line creation with Bèzier segments throwing "Command of type editLineSegment needs to prepare undo/redo info but did not.".
   * THFile.copyWith() does not deep copy elements.
+  * Repetitive line simplification isn´t written in file.
 * Infrastructure maintenance:
   * THCommmandOption changed from having its parent to having only its parent MPID.
   * MPCommand migration to _prepareUndoRedoInfo() completed.